refactor attribute assignments
