+++ features/agreement/agreement-activate.feature
@@ -1,44 +1,49 @@
 @agreement_activate
 Feature: Attivazione richiesta di fruizione
 
-  @agreement_activate1
+  @agreement_activate1 @resource_intensive
   Scenario Outline: Per una richiesta di fruizione precedentemente creata da un fruitore, la quale è in stato PENDING (prima attivazione), con tutti gli attributi richiesti certificati, tutti gli attributi richiesti dichiarati dal fruitore, e tutti gli attributi richiesti verificati dall’erogatore, alla richiesta di attivazione da parte di un utente con sufficienti permessi dell’ente erogatore, va a buon fine
     Given l'utente è un "<ruolo>" di "<enteErogatore>"
-    Given "PA2" possiede un attributo certificato
-    Given "PA2" possiede un attributo dichiarato
-    Given "PA2" crea un attributo verificato
+    Given "<enteCertificatore>" ha creato un attributo certificato e lo ha assegnato a "<enteFruitore>"
+    Given "<enteFruitore>" dichiara un attributo dichiarato
+    Given "<enteFruitore>" crea un attributo verificato
     Given un "admin" di "<enteErogatore>" ha già creato un e-service in stato "PUBLISHED" che richiede quegli attributi con approvazione "MANUAL"
-    Given "PA2" ha una richiesta di fruizione in stato "PENDING" per quell'e-service
-    Given "<enteErogatore>" verifica l'attributo verificato a "PA2"
+    Given "<enteFruitore>" ha una richiesta di fruizione in stato "PENDING" per quell'e-service
+    Given "<enteErogatore>" verifica l'attributo verificato a "<enteFruitore>"
     When l'utente richiede una operazione di attivazione di quella richiesta di fruizione
     Then si ottiene status code <risultato>
 
     Examples: 
-      | enteErogatore | ruolo        | risultato |
-      | GSP           | admin        |       200 |
-      | GSP           | api          |       403 |
-      | GSP           | security     |       403 |
-      | GSP           | support      |       403 |
-      | GSP           | api,security |       403 |
-      | PA1           | admin        |       200 |
-      | PA1           | api          |       403 |
-      | PA1           | security     |       403 |
-      | PA1           | support      |       403 |
-      | PA1           | api,security |       403 |
+      | enteFruitore | enteCertificatore | enteErogatore | ruolo        | risultato |
+      | PA1          | PA2               | GSP           | admin        |       200 |
+      | PA1          | PA2               | GSP           | api          |       403 |
+      | PA1          | PA2               | GSP           | security     |       403 |
+      | PA1          | PA2               | GSP           | support      |       403 |
+      | PA1          | PA2               | GSP           | api,security |       403 |
+      | GSP          | PA2               | PA1           | admin        |       200 |
+      | GSP          | PA2               | PA1           | api          |       403 |
+      | GSP          | PA2               | PA1           | security     |       403 |
+      | GSP          | PA2               | PA1           | support      |       403 |
+      | GSP          | PA2               | PA1           | api,security |       403 |
 
   @agreement_activate2
   Scenario Outline: Per una richiesta di fruizione precedentemente creata da un fruitore, la quale è in stato SUSPENDED (riattivazione), con tutti gli attributi richiesti certificati, tutti gli attributi richiesti dichiarati dal fruitore, e tutti gli attributi richiesti verificati dall’erogatore, alla richiesta di attivazione da parte di un utente con sufficienti permessi dell’ente erogatore, va a buon fine.
-    Given l'utente è un "admin" di "PA1"
-    Given "PA2" possiede un attributo certificato
-    Given "PA2" possiede un attributo dichiarato
-    Given "PA2" crea un attributo verificato
-    Given un "admin" di "PA1" ha già creato un e-service in stato "PUBLISHED" che richiede quegli attributi con approvazione "AUTOMATIC"
-    Given "PA2" ha una richiesta di fruizione in stato "PENDING" per quell'e-service
-    Given "PA1" verifica l'attributo verificato a "PA2"
-    Given "PA1" ha già approvato quella richiesta di fruizione
-    Given "PA1" ha già sospeso quella richiesta di fruizione come "PRODUCER"
+    Given l'utente è un "admin" di "<enteErogatore>"
+    Given "<enteCertificatore>" ha creato un attributo certificato e lo ha assegnato a "<enteFruitore>"
+    Given "<enteFruitore>" dichiara un attributo dichiarato
+    Given "<enteFruitore>" crea un attributo verificato
+    Given un "admin" di "<enteErogatore>" ha già creato un e-service in stato "PUBLISHED" che richiede quegli attributi con approvazione "AUTOMATIC"
+    Given "<enteFruitore>" ha una richiesta di fruizione in stato "PENDING" per quell'e-service
+    Given "<enteErogatore>" verifica l'attributo verificato a "<enteFruitore>"
+    Given "<enteErogatore>" ha già approvato quella richiesta di fruizione
+    Given "<enteErogatore>" ha già sospeso quella richiesta di fruizione come "PRODUCER"
+    Given "<enteFruitore>" ha già sospeso quella richiesta di fruizione come "CONSUMER"
     When l'utente richiede una operazione di attivazione di quella richiesta di fruizione
     Then si ottiene status code 200
+
+    Examples: 
+      | enteFruitore | enteCertificatore | enteErogatore |
+      | PA1          | PA2               | GSP           |
 
   @agreement_activate3
   Scenario Outline: Per una richiesta di fruizione precedentemente creata da un fruitore, la quale è in stato PENDING o SUSPENDED; con tutti gli attributi richiesti certificati, i quali sono due gruppi di due, dei quali il fruitore ne possiede uno per gruppo; tutti gli attributi richiesti dichiarati dal fruitore, i quali sono due gruppi di due, dei quali il fruitore ne possiede uno per gruppo; tutti gli attributi richiesti verificati dall’erogatore, i quali sono due gruppi di due, dei quali il fruitore ne possiede uno per gruppo; alla richiesta di attivazione da parte di un utente con sufficienti permessi dell’ente erogatore, va a buon fine.
