+++ features/agreement/step_definitions/agreement-common-steps.ts
@@ -2,7 +2,11 @@ import { Given } from "@cucumber/cucumber";
 import { z } from "zod";
 import { Role, TenantType } from "../../common-steps";
 import { dataPreparationService } from "../../../services/data-preparation.service";
-import { assertContextSchema, getToken } from "../../../utils/commons";
+import {
+  assertContextSchema,
+  getOrganizationId,
+  getToken,
+} from "../../../utils/commons";
 import {
   EServiceDescriptorState,
   AgreementApprovalPolicy,
@@ -75,5 +79,69 @@ Given(
         this.eserviceId,
         this.descriptorId
       );
+  }
+);
+
+Given(
+  "{string} ha una richiesta di fruizione in stato {string} per ognuno di quegli e-services",
+  async function (consumer: TenantType, agreementState: string) {
+    assertContextSchema(this, {
+      publishedEservicesIds: z.array(z.array(z.string())),
+      token: z.string(),
+    });
+    const token = getToken(this.tokens, consumer, "admin");
+
+    this.agreementIds = await Promise.all(
+      this.publishedEservicesIds.map(([eserviceId, descriptorId]) =>
+        dataPreparationService.createAgreementWithGivenState(
+          token,
+          agreementState,
+          eserviceId,
+          descriptorId
+        )
+      )
+    );
+  }
+);
+
+Given(
+  "{string} ha creato un attributo certificato e lo ha assegnato a {string}",
+  async function (certifier: TenantType, tenantType: TenantType) {
+    assertContextSchema(this);
+    const token = getToken(this.tokens, certifier, "admin");
+
+    const tenantId = getOrganizationId(tenantType);
+    this.attributeId = await dataPreparationService.createAttribute(
+      token,
+      "CERTIFIED"
+    );
+
+    await dataPreparationService.assignCertifiedAttributeToTenant(
+      token,
+      tenantId,
+      this.attributeId
+    );
+  }
+);
+
+Given(
+  "{string} dichiara un attributo dichiarato",
+  async function (tenantType: TenantType) {
+    assertContextSchema(this);
+    const tenantId = getOrganizationId(tenantType);
+    const token = getToken(this.tokens, tenantType, "admin");
+
+    const attributeId = await dataPreparationService.createAttribute(
+      token,
+      "DECLARED"
+    );
+
+    await dataPreparationService.declareDeclaredAttribute(
+      token,
+      tenantId,
+      attributeId
+    );
+
+    this.requiredDeclaredAttributes = [[attributeId]];
   }
 );
